Remove the Autoloader in favour of logic in the `ProphecyTrait.php` file
Includes renaming the file in which the _real_ trait lives to `ProphecyTraitReal.php`.
